More stringify cleanup
