+++ src/atlas.rs
@@ -7,12 +7,9 @@ mod crash;
 pub use crash::on_crash;
 
 mod paths;
-pub use paths::get_exe_folder;
+pub use paths::{get_exe_folder, EasyPath};
 
 mod point;
 pub use point::{Point, SizedPoint};
-
-mod easy_path;
-pub use easy_path::EasyPath;
 
 pub type BoxResult<T> = Result<T, Box<dyn std::error::Error>>;
